Re-implement functionality for saving tasks
