+++ src/data.txt
@@ -1,0 +1,3 @@
+T|0|1
+E|0|event|1|3
+D|1|d|tonight!!!!!!!!!
